requirements.txt
click==7.1.2
gitdb==4.0.5
GitPython==3.1.12
passpy==1.0.2
pyperclip==1.8.1
python-dotenv==0.15.0
python-gnupg==0.4.6
smmap==3.0.4
xxtea==2.0.0.post0
